refactor(registration-react): Implement ApiRegistrationPage as composition of independent enhancers
