--- applications/registration-react/src/pages/api-registration-page/connected-api-registration-page.jsx
@@ -1,7 +1,0 @@
-import { apiRegistrationConnector } from './api-registration-connector';
-import { APIRegistrationPage } from './api-registration-page';
-import { apiRegistrationResolver } from './api-registration-resolver';
-
-export const ConnectedAPIRegistrationPage = apiRegistrationConnector(
-  apiRegistrationResolver(APIRegistrationPage)
-);
